Update DatabaseConnector.java
Added methods for getting char info for the character select screen, creating accounts, creating characters, and getting the accountID for a user account.  The create char method does NOT check that the char name is unique - it is more of an insert a row to the char table method, at the moment.