Update DatabaseConnector.java
Updated some error checking in a few methods.  Also modified create char to also make a row in the inventory table.  Added fields for abilities.  This code is untested and the database needs to be modified to reflect the code that this contains.  We need to add an abilities field and a cooldown field to the charInfoTable.