Update Main.java
Added a few commented out calls to db methods to test SQL injection attacks.  The deleteCharacter() call uses a precisely constructed string that causes destructive SQL to be formed.  However, when that sql was called from the java the table was not dropped.  Copying and pasting the same sql into the phpmyadmin tool and then executing the script did result in the table being dropped.  Further testing of the java framework will be necessary before an exact solution can be decided upon.