Update DatabaseConnector.java
I attempted to fix the issue of the server crashing when numerous people are very active and someone logs in.  I was able to log in with two characters at the same time, but to really test these changes we need a couple of people actively playing and another person to try to log in.