Update DatabaseConnector.java
Made changes so that the password is no longer stored as plaintext.  Now, the password and the salt field are hashed, and the resulting hash is stored.  The createAccount method makes a random number for salt.  Every password request is now a result of concatenating the salt + password and then performing a SHA256 digest of this data.